@@ -1,19 +1,55 @@
 {
-    "scripts": {
-        "lint:fix": "prettier */*.js \"*/**/*{.js,.ts}\" -w",
-        "lint": "prettier */*.js \"*/**/*{.js,.ts}\" --check"
-    },
-    "dependencies": {
-        "@project-serum/anchor": "^0.24.2"
-    },
-    "devDependencies": {
-        "chai": "^4.3.4",
-        "mocha": "^9.0.3",
-        "ts-mocha": "^8.0.0",
-        "@types/bn.js": "^5.1.0",
-        "@types/chai": "^4.3.0",
-        "@types/mocha": "^9.0.0",
-        "typescript": "^4.3.5",
-        "prettier": "^2.6.2"
-    }
+  "scripts": {
+    "lint:fix": "prettier */*.js \"*/**/*{.js,.ts,.json}\" -w",
+    "lint": "prettier */*.js \"*/**/*{.js,.ts,.json}\" --check",
+    "app:dev": "next dev",
+    "app:build": "yarn anchor:test && next build",
+    "app:start": "next start",
+    "anchor:type": "cp ./target/types/*.ts ./src/types/contract.ts",
+    "anchor:idl": "cp ./target/idl/*.json ./src/configs/contract.json",
+    "anchor:build": "anchor build && yarn anchor:type && yarn lint:fix",
+    "anchor:test": "yarn anchor:build && anchor test && yarn anchor:idl"
+  },
+  "dependencies": {
+    "@animxyz/react": "^0.6.6",
+    "@project-serum/anchor": "^0.24.2",
+    "@saintno/needed-tools": "^0.1.1",
+    "@solana/wallet-adapter-base": "^0.9.5",
+    "@solana/wallet-adapter-react": "^0.15.5",
+    "@solana/wallet-adapter-react-ui": "^0.9.7",
+    "@solana/wallet-adapter-wallets": "^0.16.1",
+    "@solana/web3.js": "^1.43.1",
+    "animate.css": "^4.1.1",
+    "classnames": "^2.3.1",
+    "intercept-stdout": "^0.1.2",
+    "next": "^12.1.6",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-transition-group": "^4.4.2",
+    "recoil": "^0.7.3-alpha.2",
+    "sass": "^1.52.1"
+  },
+  "devDependencies": {
+    "@types/expect": "^24.3.0",
+    "@types/intercept-stdout": "^0.1.0",
+    "@types/mocha": "^9.1.1",
+    "@types/react": "^18.0.9",
+    "@types/react-transition-group": "^4.4.4",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "autoprefixer": "^10.4.7",
+    "chai": "^4.3.4",
+    "eslint": "8.14.0",
+    "eslint-config-next": "12.1.5",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "mocha": "^9.0.3",
+    "postcss": "^8.4.14",
+    "prettier": "^2.6.2",
+    "tailwindcss": "^3.0.24",
+    "ts-mocha": "^10.0.0",
+    "typescript": "^4.6.4"
+  }
 }
