@@ -2,9 +2,9 @@
 // single deploy script that's invoked from the CLI, injecting a provider
 // configured from the workspace's Anchor.toml.
 
-const anchor = require("@project-serum/anchor");
+import * as anchor from '@project-serum/anchor';
 
-module.exports = async function (provider) {
+module.exports = async function (provider: anchor.Provider) {
   // Configure client to use the provider.
   anchor.setProvider(provider);
 
